Created modules class in modules.py to perform model creation. train.py will only take a module class to train its model. Removed never-used constants from train.py and modules.py
